exception files reorganisation again
