+++ backend/src/main/java/marketplace/backend/exception/exceptions/user/UserNotFoundException.java
@@ -1,4 +1,4 @@
-package marketplace.backend.exception.user.exceptions;
+package marketplace.backend.exception.exceptions.user;
 
 public class UserNotFoundException extends RuntimeException {
